pyproject.toml
[project]
name = "MKSC-RL"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gymnasium==1.2.0",
    "opencv-python==4.12.0.88",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.1.1",
    "pygame==2.6.1",
    "rl-zoo3>=2.7.0",
    "setuptools<81",
    "stable-baselines3>=2.7.1",
    "swig>=4.4.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tensorboard>=2.20.0",
]
